Add a Route53 record set identifier where needed
gh-305

The last build failed with:

Invalid request: Missing field 'SetIdentifier' in Change with
[Action=CREATE, Name=305-weight.dougnaphas.com., Type=A,
SetIdentifier=null]

The docs

https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-route53-recordset.html

say

SetIdentifier
Resource record sets that have a routing policy other than simple: An
identifier that differentiates among multiple resource record sets that
have the same combination of name and type, such as multiple weighted
resource record sets named acme.example.com that have a type of A. In a
group of resource record sets that have the same name and type, the
value of SetIdentifier must be unique for each resource record set.

This is my case. I am using a weighted record set.